refactoring services structure (#15)
