+++ ingest/base_service.go
@@ -1,4 +1,4 @@
-package service
+package ingest
 
 import (
 	"errors"
