+++ ingest/ingest.go
@@ -5,53 +5,52 @@ import (
 	"log/slog"
 
 	"github.com/cometbft/rpc-companion/config"
-	"github.com/cometbft/rpc-companion/service/base"
-	"github.com/cometbft/rpc-companion/service/ingest/fetcher"
 )
 
-type Service struct {
-	service.BaseService
+// IngestService orchestrates the ingest services
+type IngestService struct {
+	BaseService
 	config  *config.Config
-	fetcher *fetcher.Fetcher
+	fetcher *Fetcher
 	//storage storage.IStorage
 }
 
 func NewIngestService(
 	logger slog.Logger,
 	config config.Config,
-) (*Service, error) {
+) (*IngestService, error) {
 
 	// Instantiate new fetcher (gRPC client)
-	fetcher, err := fetcher.NewFetcher(logger, &config)
+	fetcher, err := NewFetcher(logger, &config)
 	if err != nil {
 		return nil, fmt.Errorf("error creating new fetcher: %v", err)
 	}
 
 	// Configure Fetcher service
-	fetcher.BaseService = *service.NewBaseService(logger, "Fetcher", fetcher)
+	fetcher.BaseService = *NewBaseService(logger, "Fetcher", fetcher)
 	fetcher.SetLogger(*logger.With("service", "fetcher"))
 
 	// Ingest Service
-	ingest := &Service{
+	ingest := &IngestService{
 		config:  &config,
 		fetcher: fetcher,
 		//storage: &db,
 	}
 
-	ingest.BaseService = *service.NewBaseService(logger, "Ingest", ingest)
+	ingest.BaseService = *NewBaseService(logger, "Ingest", ingest)
 	ingest.SetLogger(*logger.With("service", "ingest"))
 
 	return ingest, nil
 }
 
-func (s *Service) OnStart() error {
+func (s *IngestService) OnStart() error {
 	if s.IsRunning() {
 		s.fetcher.Start()
 	}
 	return nil
 }
 
-func (s *Service) OnStop() {
+func (s *IngestService) OnStop() {
 	if s.fetcher.IsRunning() {
 		s.fetcher.Stop()
 	}
